adding default activity check
